refactor: add logging extension module for log initialization
